@@ -24,19 +24,19 @@ namespace Xbbcode\Tag;
 
 /**
  * Class Nobb
- * Класс для тега [nobb]
+ * Класс для тега [nobb].
  */
 class Nobb extends Tag
 {
-    const BEHAVIOUR = 'code';
+    public const BEHAVIOUR = 'code';
 
     /**
-     * Return html code
+     * Return html code.
      *
      * @return string
      */
     public function __toString()
     {
-        return \htmlspecialchars($this->getTreeText(), ENT_NOQUOTES);
+        return \htmlspecialchars($this->getTreeText(), \ENT_NOQUOTES);
     }
 }
