@@ -1,4 +1,5 @@
 <?php
+
 namespace Xbbcode\Tests\Tag;
 
 use Xbbcode\Xbbcode;
@@ -13,13 +14,12 @@ class ATest extends \PHPUnit\Framework\TestCase
         $xbbcode = new Xbbcode();
         $xbbcode->setAutoLinks(true);
         $xbbcode->parse($text);
-        $this->assertEquals($result, $xbbcode->getHtml());
+        self::assertEquals($result, $xbbcode->getHtml());
 
         $xbbcode = new Xbbcode();
         $xbbcode->setAutoLinks(false);
         $xbbcode->parse($text);
-        $this->assertEquals($result, $xbbcode->getHtml());
-
+        self::assertEquals($result, $xbbcode->getHtml());
 
         $text = 'test [url]https://github.com/Gemorroj/xBBCode#привет[/url].';
         $result = 'test <a class="bb" href="https://github.com/Gemorroj/xBBCode#%D0%BF%D1%80%D0%B8%D0%B2%D0%B5%D1%82">https://github.com/Gemorroj/xBBCode#привет</a>.';
@@ -27,11 +27,11 @@ class ATest extends \PHPUnit\Framework\TestCase
         $xbbcode = new Xbbcode();
         $xbbcode->setAutoLinks(true);
         $xbbcode->parse($text);
-        $this->assertEquals($result, $xbbcode->getHtml());
+        self::assertEquals($result, $xbbcode->getHtml());
 
         $xbbcode = new Xbbcode();
         $xbbcode->setAutoLinks(false);
         $xbbcode->parse($text);
-        $this->assertEquals($result, $xbbcode->getHtml());
+        self::assertEquals($result, $xbbcode->getHtml());
     }
 }
