@@ -1,4 +1,5 @@
 <?php
+
 namespace Xbbcode\Tests\Tag;
 
 use Xbbcode\Xbbcode;
@@ -12,6 +13,6 @@ class AbbrTest extends \PHPUnit\Framework\TestCase
 
         $xbbcode = new Xbbcode();
         $xbbcode->parse($text);
-        $this->assertEquals($result, $xbbcode->getHtml());
+        self::assertEquals($result, $xbbcode->getHtml());
     }
 }
