@@ -1,4 +1,5 @@
 <?php
+
 namespace Xbbcode\Tests\Tag;
 
 use Xbbcode\Xbbcode;
@@ -12,14 +13,13 @@ class GoogleTest extends \PHPUnit\Framework\TestCase
 
         $xbbcode = new Xbbcode();
         $xbbcode->parse($text);
-        $this->assertEquals($result, $xbbcode->getHtml());
-
+        self::assertEquals($result, $xbbcode->getHtml());
 
         $text = 'test [google=xBBCode]search[/google].';
         $result = 'test <a class="bb" href="//www.google.com/search?q=xBBCode">search</a>.';
 
         $xbbcode = new Xbbcode();
         $xbbcode->parse($text);
-        $this->assertEquals($result, $xbbcode->getHtml());
+        self::assertEquals($result, $xbbcode->getHtml());
     }
 }
