@@ -1,4 +1,5 @@
 <?php
+
 namespace Xbbcode\Tests\Tag;
 
 use Xbbcode\Xbbcode;
@@ -12,14 +13,13 @@ class YandexTest extends \PHPUnit\Framework\TestCase
 
         $xbbcode = new Xbbcode();
         $xbbcode->parse($text);
-        $this->assertEquals($result, $xbbcode->getHtml());
-
+        self::assertEquals($result, $xbbcode->getHtml());
 
         $text = 'test [yandex=xBBCode]search[/yandex].';
         $result = 'test <a class="bb" href="//yandex.com/yandsearch?text=xBBCode">search</a>.';
 
         $xbbcode = new Xbbcode();
         $xbbcode->parse($text);
-        $this->assertEquals($result, $xbbcode->getHtml());
+        self::assertEquals($result, $xbbcode->getHtml());
     }
 }
